+++ src/db/schema.sql
@@ -334,8 +334,12 @@ CREATE TABLE IF NOT EXISTS schema_meta (
 -- are pruned opportunistically, so this table stays small.
 -- ---------------------------------------------------------------------------
 CREATE TABLE IF NOT EXISTS rate_hits (
-  key TEXT    NOT NULL,
-  ts  INTEGER NOT NULL
+  key TEXT   NOT NULL,
+  -- BIGINT, not INTEGER: this holds Date.now() in milliseconds, which passed
+  -- 2^31 in 2001. SQLite's INTEGER is 8 bytes and hides that, but Postgres
+  -- INTEGER is 4 and rejects the value outright, taking every rate-limited
+  -- endpoint down with it.
+  ts  BIGINT NOT NULL
 );
 
 CREATE INDEX IF NOT EXISTS idx_rate_hits_key_ts ON rate_hits (key, ts);
